requirements.txt
fastapi~=0.104.1
pydantic==2.5.1
python-jose==3.3.0
python-multipart==0.0.6
sqlalchemy==2.0.23
uvicorn~=0.24.0.post1
psycopg2==2.9.9
